ocu user do not login app

+++ app/controllers/devise/oauth2_providable/tokens_controller.rb
@@ -4,13 +4,19 @@ class Devise::Oauth2Providable::TokensController < ApplicationController
   include UsersHelper
 
   def create
+    binding.pry
     if oauth2_current_refresh_token then
       @refresh_token = oauth2_current_refresh_token
     else
 
       if oauth2_current_client then
-        Devise::Oauth2Providable::RefreshToken.where(:account_id=>current_account.id,:client_id=>[1,2]).delete_all
-        @refresh_token = oauth2_current_client.refresh_tokens.create!(:account_id => current_account.id)
+        # Ocu user do not allow to login app and website.
+        if current_account.home_user.role_code == 3 and oauth2_current_client.id != ::Const::OCU_PLATFORM_APP then
+            return render(:json => {:errors=>{:message=>"公众号不能登录客户端.",:status_code=>:invalid_request}},:status => 403)
+        else
+            Devise::Oauth2Providable::RefreshToken.where(:account_id=>current_account.id,:client_id=>[1,2]).delete_all
+            @refresh_token = oauth2_current_client.refresh_tokens.create!(:account_id => current_account.id)
+        end
       else
 
         Rails.logger.error "Oauth2 create token error: Oauth client not found, can your see current_account id:#{current_account.id}"
